Reverify public downloads nightly
