+++ Tests/QuillCodeParityTests/ParityDownloadBuildsGateTests.swift
@@ -316,6 +316,15 @@ final class ParityDownloadBuildsGateTests: QuillCodeParityTestCase {
             "scripts/plan-download-build.sh",
             "build-required: ${{ steps.plan.outputs.build-required }}",
             "if: needs.release-policy.outputs.build-required == 'true'",
+            "verify-current-published:",
+            "if: needs.release-policy.outputs.build-required == 'false'",
+            "Reverify current public tester release",
+            "--discover-workflow-run-url",
+            "verify-current-daily-driver:",
+            "needs: [release-policy, verify-current-published]",
+            "scripts/public-macos-daily-driver-smoke.sh",
+            "name: quillcode-public-daily-driver-${{ matrix.arch }}",
+            "retention-days: 30",
             "persist-credentials: false",
             "needs: release-policy",
             "capture-updater-source:",
@@ -458,6 +467,9 @@ final class ParityDownloadBuildsGateTests: QuillCodeParityTestCase {
             "returns the new release to draft",
             "previous stable feed",
             "avoids no-op build-number updates and unnecessary",
+            "does not skip product verification",
+            "three-process 100-chat daily-driver workload",
+            "Raw reports and screenshots are retained for 30 days",
             "untouched previous public app",
             "synthetic one-build-behind fallback",
             "channel is `tester`",
