+++ scripts/verify-published-release.py
@@ -4,15 +4,10 @@
 from __future__ import annotations
 
 import argparse
-import hashlib
 import os
 import tempfile
-import time
-import urllib.error
-import urllib.request
-from collections.abc import Callable
 from pathlib import Path
-from typing import Any, TypeVar
+from typing import Any
 from urllib.parse import quote
 
 from release_verification_contract import (
@@ -21,14 +16,19 @@ from release_verification_contract import (
     REPOSITORY_PATTERN,
     SHA_PATTERN,
     VerificationError,
+    discover_workflow_run_url,
     expected_urls,
     load_json_bytes,
     validate_manifest,
 )
 from release_verification_files import read_bounded, verify_asset_files
-
-
-Result = TypeVar("Result")
+from release_verification_remote import (
+    api_json,
+    download_asset,
+    fetch_bytes,
+    resolve_tag_commit,
+    retry,
+)
 
 
 def parse_arguments() -> argparse.Namespace:
@@ -44,10 +44,15 @@ def parse_arguments() -> argparse.Namespace:
         help="Require a public stable prerelease that has not been promoted to latest.",
     )
     parser.add_argument("--commit", required=True, help="Expected 40-character commit SHA.")
-    parser.add_argument(
+    provenance = parser.add_mutually_exclusive_group(required=True)
+    provenance.add_argument(
         "--workflow-run-url",
-        required=True,
         help="Expected publishing run URL.",
+    )
+    provenance.add_argument(
+        "--discover-workflow-run-url",
+        action="store_true",
+        help="Discover the expected publishing run URL from the bounded public manifest.",
     )
     parser.add_argument(
         "--attempts",
@@ -124,159 +129,6 @@ def parse_arguments() -> argparse.Namespace:
     return arguments
 
 
-def fetch_bytes(
-    url: str,
-    byte_limit: int,
-    *,
-    token: str | None = None,
-    accept: str = "application/octet-stream",
-) -> bytes:
-    headers = {
-        "Accept": accept,
-        "User-Agent": "Quill-Cowork-Release-Verifier/1",
-        "X-GitHub-Api-Version": "2022-11-28",
-    }
-    if token:
-        headers["Authorization"] = f"Bearer {token}"
-    request = urllib.request.Request(url, headers=headers)
-    try:
-        with urllib.request.urlopen(request, timeout=30) as response:
-            if response.status != 200:
-                raise VerificationError(f"GET {url} returned HTTP {response.status}")
-            length = response.headers.get("Content-Length")
-            if length and int(length) > byte_limit:
-                raise VerificationError(
-                    f"GET {url} exceeds its {byte_limit}-byte limit"
-                )
-            data = response.read(byte_limit + 1)
-    except VerificationError:
-        raise
-    except urllib.error.HTTPError as error:
-        raise VerificationError(f"GET {url} returned HTTP {error.code}") from error
-    except urllib.error.URLError as error:
-        raise VerificationError(f"GET {url} failed: {error.reason}") from error
-    except (OSError, ValueError) as error:
-        raise VerificationError(f"GET {url} failed: {error}") from error
-    if len(data) > byte_limit:
-        raise VerificationError(f"GET {url} exceeds its {byte_limit}-byte limit")
-    return data
-
-
-def api_json(repo: str, path: str, token: str | None) -> dict[str, Any]:
-    url = f"https://api.github.com/repos/{repo}/{path}"
-    return load_json_bytes(
-        fetch_bytes(
-            url,
-            API_BYTE_LIMIT,
-            token=token,
-            accept="application/vnd.github+json",
-        ),
-        url,
-    )
-
-
-def resolve_tag_commit(repo: str, tag: str, token: str | None) -> str:
-    reference = api_json(repo, f"git/ref/tags/{quote(tag, safe='')}", token)
-    target = reference.get("object")
-    for _ in range(5):
-        if not isinstance(target, dict):
-            break
-        target_type = target.get("type")
-        sha = target.get("sha")
-        if not isinstance(sha, str) or not SHA_PATTERN.fullmatch(sha):
-            break
-        if target_type == "commit":
-            return sha
-        if target_type != "tag":
-            break
-        tag_object = api_json(repo, f"git/tags/{sha}", token)
-        target = tag_object.get("object")
-    raise VerificationError(f"release tag {tag!r} does not resolve to a commit")
-
-
-def download_asset(url: str, destination: Path, size: int, digest: str) -> None:
-    request = urllib.request.Request(
-        url,
-        headers={
-            "Accept": "application/octet-stream",
-            "User-Agent": "Quill-Cowork-Release-Verifier/1",
-        },
-    )
-    temporary = destination.with_suffix(destination.suffix + ".part")
-    try:
-        with urllib.request.urlopen(request, timeout=60) as response, temporary.open(
-            "xb"
-        ) as output:
-            if response.status != 200 or not response.geturl().lower().startswith(
-                "https://"
-            ):
-                raise VerificationError(
-                    f"download for {destination.name!r} returned an invalid response"
-                )
-            response_length = response.headers.get("Content-Length")
-            if response_length and int(response_length) != size:
-                raise VerificationError(
-                    f"download for {destination.name!r} reported the wrong size"
-                )
-            hasher = hashlib.sha256()
-            downloaded = 0
-            while chunk := response.read(1024 * 1024):
-                downloaded += len(chunk)
-                if downloaded > size:
-                    raise VerificationError(
-                        f"download for {destination.name!r} exceeded its size"
-                    )
-                output.write(chunk)
-                hasher.update(chunk)
-        if downloaded != size or hasher.hexdigest() != digest:
-            raise VerificationError(
-                f"download for {destination.name!r} failed integrity verification"
-            )
-        temporary.replace(destination)
-    except VerificationError:
-        temporary.unlink(missing_ok=True)
-        raise
-    except urllib.error.HTTPError as error:
-        temporary.unlink(missing_ok=True)
-        raise VerificationError(
-            f"download for {destination.name!r} returned HTTP {error.code}"
-        ) from error
-    except urllib.error.URLError as error:
-        temporary.unlink(missing_ok=True)
-        raise VerificationError(
-            f"download for {destination.name!r} failed: {error.reason}"
-        ) from error
-    except (OSError, ValueError) as error:
-        temporary.unlink(missing_ok=True)
-        raise VerificationError(
-            f"download for {destination.name!r} failed: {error}"
-        ) from error
-
-
-def retry(
-    operation: Callable[[], Result],
-    *,
-    attempts: int,
-    delay: float,
-    label: str,
-) -> Result:
-    last_error: VerificationError | None = None
-    for attempt in range(1, attempts + 1):
-        try:
-            return operation()
-        except VerificationError as error:
-            last_error = error
-            if attempt < attempts:
-                print(
-                    f"{label} attempt {attempt} failed: {error}; retrying",
-                    flush=True,
-                )
-                time.sleep(delay)
-    if last_error is None:
-        raise VerificationError(f"{label} did not run")
-    raise last_error
-
-
 def validate_snapshot(
     release: dict[str, Any],
     manifest_bytes: bytes,
@@ -334,6 +186,11 @@ def verify_offline(arguments: argparse.Namespace) -> None:
         MANIFEST_BYTE_LIMIT,
         "manifest",
     )
+    if arguments.discover_workflow_run_url:
+        arguments.workflow_run_url = discover_workflow_run_url(
+            manifest_bytes,
+            arguments.repo,
+        )
     manifest, assets = validate_snapshot(
         release,
         manifest_bytes,
@@ -373,6 +230,16 @@ def verify_offline(arguments: argparse.Namespace) -> None:
 def verify_public(arguments: argparse.Namespace) -> None:
     token = os.environ.get("GH_TOKEN") or os.environ.get("GITHUB_TOKEN")
     urls = expected_urls(arguments.repo, arguments.tag, arguments.channel)
+    if arguments.discover_workflow_run_url:
+        arguments.workflow_run_url = retry(
+            lambda: discover_workflow_run_url(
+                fetch_bytes(urls["manifest"], MANIFEST_BYTE_LIMIT),
+                arguments.repo,
+            ),
+            attempts=arguments.attempts,
+            delay=arguments.retry_delay,
+            label="public release provenance",
+        )
 
     def fetch_snapshot() -> tuple[dict[str, Any], list[dict[str, Any]]]:
         release = api_json(
